% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bk.dat}
\alias{bk.dat}
\title{Title}
\description{
RNA-seq sequencing data with count values, where column names are gene symbols and row names are sample IDs.

Survival/Clinical data, consisting of three columns, with the first column representing time, the second column representing status and the laste column representing sample types.consisting of one column with row names corresponding to sample names.

Clinical data, consisting of three columns with row names corresponding to sample names.
}
