% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Melanocyte_proportion.R
\name{Melanocyte_proportion}
\alias{Melanocyte_proportion}
\title{Title}
\usage{
Melanocyte_proportion(
  bk.dat = NULL,
  out_dir = NULL,
  key = "Melanocyte_cluster4"
)
}
\arguments{
\item{bk.dat}{RNA-seq sequencing data with count values, where column names are gene symbols and row names are sample IDs.}

\item{out_dir}{Output results folder.}

\item{key}{The character string that correponds to malignant cells in cell.type.labels. Set to NULL if there are no malignant cells or the malignant cells between reference and mixture are from matched sample, in which case all cell types will be treated equally, this parameter was derived from the 'new.prism' function in BayesPrism package.}
}
\value{
data.frame
}
\description{
Title
}
