% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Melanocyte_proportion_survival.R
\name{Melanocyte_proportion_survival}
\alias{Melanocyte_proportion_survival}
\title{Title}
\usage{
Melanocyte_proportion_survival(out_dir = NULL, theta = NULL, Survival = NULL)
}
\arguments{
\item{out_dir}{Output results folder.}

\item{theta}{The cell proportion matrix calculated from the 'Melanocyte_proportion' function.}

\item{Survival}{Survival data, consisting of two columns, with the first column representing time and the second column representing status.}
}
\value{
list
}
\description{
Title
}
